add sorting usage
