+++ piperider_cli/__init__.py
@@ -4,7 +4,6 @@ import os
 import re
 import sys
 import webbrowser
-from collections import deque
 from datetime import datetime
 
 from dateutil import tz
@@ -262,36 +261,3 @@ def get_run_json_path(output_dir: str, input=None):
                     latest_report_dir = de.path
             run_json = os.path.join(output_dir, latest_report_dir, 'run.json')
     return os.path.abspath(run_json)
-
-
-def topological_sort(graph, num_of_vertices):
-    in_degree = {i: 0 for i in graph}
-
-    for i in graph:
-        for j in graph[i]:
-            in_degree[j] += 1
-
-    queue = deque()
-    for i in graph:
-        if in_degree[i] == 0:
-            queue.append(i)
-
-    cnt = 0
-    top_order = []
-
-    while queue:
-        u = queue.popleft()
-        top_order.append(u)
-
-        for i in graph[u]:
-            in_degree[i] -= 1
-            if in_degree[i] == 0:
-                queue.append(i)
-
-        cnt += 1
-
-    # Check if there was a cycle
-    if cnt != num_of_vertices:
-        raise Exception("There exists a cycle in the graph. Topological sorting is not possible.")
-    else:
-        return top_order
