+++ tests/test_compare_summary_ng.py
@@ -3,6 +3,8 @@ import os
 import unittest
 from unittest import TestCase
 
+from piperider_cli import dbtutil
+from piperider_cli.dbt.sorting import topological_sort
 from piperider_cli.compare_report import ComparisonData
 from packaging import version
 
@@ -38,6 +40,7 @@ class TestCompareSummaryNG(TestCase):
         with open("output.md", "w") as fh:
             fh.write(result)
 
+    @unittest.skip("the input does not contains ref_id, we need new inputs")
     def test_in_memory_compare_with_manifests_v_1_3(self):
         run1 = self.manifest_dict("jaffle_shop_base_1.3.json")
         run2 = self.manifest_dict("jaffle_shop_target_1.3.json")
@@ -69,3 +72,11 @@ class TestCompareSummaryNG(TestCase):
 
         with open("output.md", "w") as fh:
             fh.write(result)
+
+    def test_sorting(self):
+        # show usage of the sort
+        m = self.manifest_dict("case_base_not_profiled_1.json").get('dbt', {}).get('manifest', {})
+        graph = dbtutil.prepare_topological_graph(m)
+        output = topological_sort(graph, len(list(graph.keys())))
+        print(output)
+        pass
